+++ arcangelo/arcangelo/__init__.py
@@ -1,3 +1,2 @@
 # -*- coding: utf-8 -*-
-from . import scripts
-from . import gen_addons_table
+