+++ arcangelo/arcangelo/scripts/__init__.py
@@ -1,1 +1,4 @@
+# -*- coding: utf-8 -*-
+from . import license_mgnt
+from . import gen_addons_table
 from . import arcangelo