+++ arcangelo/arcangelo/tests/data/match/odoo13/__init__.py
@@ -1,2 +1,2 @@
 # flake8: noqa - pylint: skip-file
-from . import partner
+from . import models