+++ arcangelo/arcangelo/tests/data/match/odoo8/__init__.py
@@ -1,3 +1,3 @@
 # flake8: noqa - pylint: skip-file
 # -*- coding: utf-8 -*-
-from . import partner
+from . import models