+++ arcangelo/arcangelo/egg-info/CHANGELOG.rst
@@ -1,3 +1,10 @@
+2.1.0 (2025-06-15)
+~~~~~~~~~~~~~~~~~~~
+
+* [IMP] Split from wok_code
+* [IMP] Before migration warns on different base name
+* [FIX] If target direcotry does not exist, will be create
+
 2.0.22 (2025-05-31)
 ~~~~~~~~~~~~~~~~~~~